# api/openapi.yaml
openapi: 3.1.0
info:
  title: HarborMaster API
  version: 0.2.5
  summary: Safety-first container lifecycle manager.
  description: |
    HarborMaster's REST API.

    **This API is read-only with respect to Docker.** It inventories containers,
    images, networks, and volumes, and reports build metadata and health. It
    cannot pull images, create, update, restart, or remove containers, and it
    exposes no command-execution endpoint. Those capabilities are not
    implemented yet, and will be added to this document at the same time they
    are added to the server.

    The one endpoint that accepts a write method is
    `POST /api/v1/inventory/refresh`. It mutates HarborMaster's own records --
    re-reading the host and replacing what HarborMaster has stored -- and
    changes nothing on the Docker host.

    ### Operating the API safely

    HarborMaster talks to the Docker socket, which is a privileged interface:
    anything able to reach it can control every container on the host. The
    server therefore binds to `127.0.0.1` by default. Expose it beyond loopback
    only behind an authenticating reverse proxy.

    ### Authentication

    None yet. The API is unauthenticated and is expected to be reachable only
    from loopback or a trusted proxy. Authentication is planned and will be
    documented here when it lands.

    ### Sensitive values

    Environment variables and log-driver options whose NAME matches a
...

...
    unredacted value is ever stored.

...

...

...

...

...
servers:
...
paths:
...
  /api/v1/containers:
    get:
      tags: [Containers]
      summary: List containers
      operationId: listContainers
      description: |
        Returns a page of containers from the persisted inventory, with
        filtering and sorting applied by the database.

        Ordering is always deterministic: the requested sort field is followed
        by name and then ID as tiebreaks, so paging cannot repeat or skip a
        container.

        Sort fields are validated against a fixed allowlist. Nothing from the
        query string is ever interpolated into SQL.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: search
          in: query
          description: Matches container name, image reference, or ID prefix.
          schema: { type: string }
        - name: state
          in: query
          description: |
            Filter by state. Repeat the parameter or supply a comma-separated
            list; values are OR-ed.
          schema:
            type: string
            examples: ["running", "running,paused"]
        - name: health
          in: query
          description: Filter by health, OR-ed like `state`.
          schema:
            type: string
            examples: ["unhealthy"]
        - name: project
          in: query
          description: Exact Compose project name.
          schema: { type: string }
        - name: service
          in: query
          description: Exact Compose service name.
          schema: { type: string }
        - name: image
          in: query
          description: Exact image repository or full reference.
          schema: { type: string }
        - name: restartPolicy
          in: query
          schema:
            type: string
            enum: ["no", always, on-failure, unless-stopped]
        - name: harbormasterEnabled
          in: query
          description: |
            Filter on the `io.harbormaster.enabled` label. Omitting the
            parameter does not filter; `false` matches containers where the
            label is present and false, which is not the same as absent.
          schema: { type: boolean }
        - name: labelKey
          in: query
          description: Match containers carrying this label key.
          schema: { type: string }
        - name: labelValue
          in: query
          description: Narrow `labelKey` to an exact value. Requires `labelKey`.
          schema: { type: string }
        - name: includeAbsent
          in: query
          description: |
            Include containers that earlier refreshes saw but the most recent
            one did not. Absent containers are retained rather than deleted, so
            their history survives.
          schema: { type: boolean, default: false }
        - name: sort
          in: query
          schema:
            type: string
            default: name
            enum: [name, state, health, created, started, image, project, service, restartCount, lastSeen]
        - name: direction
          in: query
          schema:
            type: string
            default: asc
            enum: [asc, desc]
      responses:
        "200":
          description: A page of containers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerList"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/InventoryUnavailable"
...
  /api/v1/images:
    get:
      tags: [Images]
      summary: List images
      operationId: listImages
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of images with reference counts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageList"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/InventoryUnavailable"
...
components:
...
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: Items per page. Values above the maximum are rejected, not clamped.
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
...
  responses:
...
    InvalidRequest:
      description: |
        A query parameter was malformed or out of range. The message names the
        parameter and what was expected; it never echoes the value supplied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: { code: invalid_request, message: pageSize must be at most 200 }
            requestId: 3f8a1c9e7b2d4a6f0c1e5b93
...
    MethodNotAllowed:
      description: |
        The path exists but the method is not permitted. Every endpoint except
        the inventory refresh is read-only.
      headers:
        Allow:
          description: Permitted methods.
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: { code: method_not_allowed, message: method not allowed }
    InventoryUnavailable:
      description: |
        The inventory is not configured or is disabled by configuration.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: { code: feature_disabled, message: inventory is disabled by configuration }
...
  schemas:
...
    Pagination:
      type: object
      required: [page, pageSize, totalItems, totalPages, hasNext, hasPrevious]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        totalItems: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasNext: { type: boolean }
        hasPrevious: { type: boolean }
    ErrorCode:
      type: string
      description: |
        Stable machine-readable identifier. Client logic must branch on this
        rather than on `message`, whose prose may change.
      enum:
        - not_found
        - method_not_allowed
        - payload_too_large
        - internal_error
        - invalid_request
        - ambiguous_id
        - conflict
        - service_unavailable
        - feature_disabled
    ErrorResponse:
      type: object
      description: The envelope returned by every non-2xx JSON response.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: |
                Short human-readable summary, safe to display. Never contains a
                stack trace, a filesystem path, an internal error chain, or a
                value the caller supplied.
        requestId:
          type: string
          description: Correlates this response with the server log.

    # ------------------------------------------------------------- system --
...
    ContainerState:
      type: string
      description: |
        Normalized lifecycle state. The vocabulary is closed; an unrecognised
        runtime state maps to `unknown` rather than passing through.
      enum: [created, running, paused, restarting, removing, exited, dead, unknown]
    HealthState:
      type: string
      description: A container with no healthcheck reports `none`.
      enum: [none, starting, healthy, unhealthy]
    ImageRef:
      type: object
      description: |
        A parsed image reference. `raw` is always the reference exactly as the
        container declared it; parsing is best-effort and never fails.
      required: [raw]
      properties:
        raw: { type: string, examples: ["nginx:1.27"] }
        repository: { type: string, examples: [nginx] }
        tag: { type: string, examples: ["1.27"] }
        digest: { type: string, examples: ["sha256:abc..."] }
    RestartPolicy:
      type: object
      required: [name]
      properties:
        name:
          type: string
          enum: ["no", always, on-failure, unless-stopped]
        maximumRetryCount:
          type: integer
          description: Applies to the on-failure policy only.
    Port:
      type: object
      required: [containerPort, protocol, published]
      properties:
        containerPort: { type: integer, minimum: 0, maximum: 65535 }
        protocol: { type: string, examples: [tcp, udp] }
        hostIp: { type: string }
        hostPort: { type: integer, minimum: 0, maximum: 65535 }
        published:
          type: boolean
          description: |
            Distinguishes a port merely exposed by the image from one actually
            bound on the host. Only published ports are reachable.
    ComposeMetadata:
      type: object
      required: [managed, oneOff]
      properties:
        managed:
          type: boolean
          description: True when the container carries a Compose project label.
        project: { type: string }
        service: { type: string }
        containerNumber: { type: integer }
        workingDir: { type: string }
        configFiles: { type: string }
        version: { type: string }
        oneOff:
          type: boolean
          description: Started by `compose run` rather than `compose up`.
    HarborMasterMetadata:
      type: object
      properties:
        enabled:
          type: boolean
          description: |
            Absent when the container carries no `io.harbormaster.enabled`
            label. Absent, true, and false are three distinct answers.
        labels:
          type: object
          description: Every `io.harbormaster.*` label, prefix stripped.
          additionalProperties: { type: string }
    ContainerSummary:
      type: object
      description: The list-row projection of a container.
      required: [hostId, id, shortId, name, image, state, health, createdAt, restartCount, restartPolicy, compose, harbormaster, ports, present, firstSeenAt, lastSeenAt, generation, warningCount]
      properties:
        hostId: { type: string }
        id: { type: string }
        shortId: { type: string }
        name: { type: string }
        image:
          $ref: "#/components/schemas/ImageRef"
        imageId: { type: string }
        state:
          $ref: "#/components/schemas/ContainerState"
        status:
          type: string
          description: Runtime status text such as "Up 3 minutes". Volatile.
        health:
          $ref: "#/components/schemas/HealthState"
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
        exitCode:
          type: integer
          description: Present only once the container has exited.
        restartCount: { type: integer, minimum: 0 }
        restartPolicy:
          $ref: "#/components/schemas/RestartPolicy"
        compose:
          $ref: "#/components/schemas/ComposeMetadata"
        harbormaster:
          $ref: "#/components/schemas/HarborMasterMetadata"
        ports:
          type: array
          items: { $ref: "#/components/schemas/Port" }
        present:
          type: boolean
          description: |
            False for a container earlier refreshes saw but the most recent one
            did not. Such rows are retained rather than deleted.
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        generation: { type: integer, format: int64 }
        warningCount: { type: integer, minimum: 0 }
    ContainerList:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ContainerSummary" }
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    Image:
      type: object
      required: [id, shortId, repoTags, repoDigests, size]
      properties:
        id: { type: string }
        shortId: { type: string }
        repoTags:
          type: array
          description: Empty for an untagged ("dangling") image.
          items: { type: string }
        repoDigests:
          type: array
          description: Usually present only for images pulled from or pushed to a registry.
          items: { type: string }
        createdAt: { type: string, format: date-time }
        architecture: { type: string }
        os: { type: string }
        osVersion: { type: string }
        variant: { type: string }
        size: { type: integer, format: int64 }
        author: { type: string }
        comment: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
    ImageUsage:
      type: object
      required: [image, containerCount]
      properties:
        image:
          $ref: "#/components/schemas/Image"
        containerCount:
          type: integer
          minimum: 0
          description: Present containers referencing this image.
    ImageList:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ImageUsage" }
        pagination:
          $ref: "#/components/schemas/Pagination"
...